+++ packages/forc-bin/README.md
@@ -1,0 +1,3 @@
+# forc-bin
+
+> Binary wrapper for [Forc](https://github.com/FuelLabs/sway)
