+++ packages/forc-bin/src/install.ts
@@ -5,27 +5,7 @@ import fetch from 'node-fetch';
 import path from 'path';
 import sh from 'shelljs';
 
-const getCurrentVersion = async () => {
-  const pkgJsonPath = path.join(__dirname, '../package.json');
-  const version = JSON.parse(await fs.readFile(pkgJsonPath, 'utf8')).version;
-  return version;
-};
-
-const getPkgPlatform = () => {
-  if (process.platform === 'darwin' && process.arch === 'arm64') {
-    return 'darwin_arm64';
-  }
-  if (process.platform === 'darwin' && process.arch === 'x64') {
-    return 'darwin_amd64';
-  }
-  if (process.platform === 'linux' && process.arch === 'arm64') {
-    return 'linux_arm64';
-  }
-  if (process.platform === 'linux' && process.arch === 'x64') {
-    return 'linux_amd64';
-  }
-  throw new Error('Unsupported platform');
-};
+import { getCurrentVersion, getPkgPlatform } from './shared';
 
 (async () => {
   const pkgPlatform = getPkgPlatform();
